# Improve `Tree` performance