@Saul2022 2024-07-27
Wouldn't this fix https://github.com/godotengine/godot/issues/94648#issue-2424447273  along with rthe closing pr?