@aaronp64 2024-07-27
> Wouldn't this fix [#94648 (comment)](https://github.com/godotengine/godot/issues/94648#issue-2424447273) along with rthe closing pr?

No, I ran into this when looking into that issue, but most of the time spent in that case is in `DynamicBVH` functions.  From their screenshots, I think the reason `Tree` wasn't an issue there is that the block nodes were internal to the map, so they weren't being added to the `SceneTreeEditor`