@AThousandShips 2024-08-28
This broke trees in at least `EditorHelpSearch`, will try to identify what was broken